Update README.md
Added a link to the readme of the Appveyor build.  Because the [repo isn't public (?)](http://www.appveyor.com/docs/status-badges#badges-for-projects-with-public-repositories-on-github-and-bitbucket) the badge doesn't show up; but the link works just fine.  If you log into AppVeyor and go to this URL:

https://ci.appveyor.com/project/jbrantly/ts-loader/settings/badges

You should be able to get the badge from there.