Emit errors of multiple typescript instances correctly
If multipe typescript instances are used, then the errors of only
one instance are outputted to the console. The error has been caused by
the removeTSLoaderErrors function that removes the errors for all
instances, therefore, only the errors of the last instance are left in
the errors array. This is fixed by including the instance name in the
loaderSource and refining removeTSLoaderErrors to only remove errors of
the same instance.

Fixes #268