Use custom transformer when building solution references (#1550)
* Refactor getting custom transformer into a dedicated function

This allows us to share the function between `initializeInstance` and
`makeSolutionBuilderHost`. `makeSolutionBuilderHost` is called as a part
of `buildSolutionReferences`.  `buildSolutionReferences` is called before
`initializeInstance` where the custom transformers are initialized.

* Implement `getCustomTransformer` in `makeSolutionBuilderHost`

This allows us to pass on the custom transformer while the project
references are built.

* Remove redundant `.toString`

`options.configFilePath` is always a string, there's no reason we need `toString()` too.

* Runtime check that project is a string

In a previous commit, I removed `.toString()` without validating what
the Typescript compiler would say. This change does a runtime check
that the project variable is an actual string.

* Update `CHANGELOG.md` and `package.json`

Co-authored-by: John Reilly <johnny_reilly@hotmail.com>